improve

+++ src/cpu/pred/statistical_corrector.hh
@@ -229,8 +229,8 @@ class StatisticalCorrector : public SimObject
         int16_t per_bank[8] = {0, 0, 0, 0, 0, 0, 0, 0}; 
         uint8_t ut_j[8] = {0, 0, 0, 0, 0, 0, 0, 0}; 
 
-        uint8_t ut_index1[8][3] = {0, 0, 0, 0, 0, 0, 0, 0};
-        uint16_t wt_index1[8][3] = {0, 0, 0, 0, 0, 0, 0, 0};
+        uint32_t ut_index1[8][3] = {0, 0, 0, 0, 0, 0, 0, 0};
+        uint32_t wt_index1[8][3] = {0, 0, 0, 0, 0, 0, 0, 0};
         uint16_t digest[32] = {0};
     };
